% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Produce a Boxplot of your variable of choice}
\usage{
boxplot(var, gender = "women")
}
\arguments{
\item{var}{variable to plot}

\item{gender}{women or men}
}
\value{

}
\description{
Produce a Boxplot of your variable of choice
}
