% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamesIsValid}
\alias{gamesIsValid}
\title{Check Games Argument Validity}
\usage{
gamesIsValid(games)
}
\arguments{
\item{games}{argument to check}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Check that games argument is valid value for ecacr use
}
\examples{
gamesIsValid("all") # TRUE
gamesIsValid("foobar") # FALSE
}
