% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genderIsValid}
\alias{genderIsValid}
\title{Checks gender is formatted correctly}
\usage{
genderIsValid(gender)
}
\arguments{
\item{gender}{gender to check}
}
\value{
TRUE if gender is formatted correctly, FALSE otherwise
}
\description{
Checks gender is formatted correctly
}
