% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Produce a ggplot2 Histogram}
\usage{
histogram(var, games = "all", players = "goalies", gender = "women")
}
\arguments{
\item{var}{variable to plot}

\item{games}{graphing 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{players}{type of players stats to graph, 'goalies' (default) or 'skaters'}

\item{gender}{'women' (default) or 'men'}
}
\value{
a ggplot2 histogram object
}
\description{
Create a ggplot2 histogram object displaying the distribution of a given statistic among
all ECAC players of a single gender (women or men)
}
\examples{
histogram(var="Saves")
histogram(var="Goals", games="conference", players="skaters", gender="men")
}
