% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listValidTeams}
\alias{listValidTeams}
\title{List of Valid Teams}
\usage{
listValidTeams()
}
\value{
a list
}
\description{
Return a list of valid team names for ecacr package
}
\examples{
listValidTeams()
}
