% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopChart.R
\name{lollipopChart}
\alias{lollipopChart}
\title{Make a Lollipop Chart}
\usage{
lollipopChart(var, gender = "women")
}
\arguments{
\item{var}{variable to graph, must be given as a string}

\item{gender}{gender of data to graph}
}
\value{
returns a lollipop chart of the given variable, separated by team. Plots Games Played by default.
}
\description{
Make a Lollipop Chart
}
\examples{
lollipopChart() returns a plot of women's teams games played
lollipopChart("Goals") returns a plot of given women's teams goals scored
lollipopChart("Goals", "men") returns a plot of men's team's goals scored
}
