% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallLeaders.R
\name{overallLeaders}
\alias{overallLeaders}
\title{Produce a Chart of Individual Overall Season Leaders}
\usage{
overallLeaders(
  var,
  n_leaders = 10,
  games = "all",
  players = "goalies",
  gender = "women",
  verbose = TRUE
)
}
\arguments{
\item{var}{summary statistic to plot}

\item{n_leaders}{number of leaders to plot (default is 10)}

\item{games}{graphing 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{players}{type of players stats to graph, 'goalies' (default) or 'skaters'}

\item{gender}{'women' (default) or 'men'}

\item{verbose}{TRUE (default) for intermediate data scraping output messages, FALSE for no additional output}
}
\value{
a ggplot2 lollipop chart object
}
\description{
Creates a ggplot2 lollipop chart (bar chart alternative) object. Displays the overall individual leaders of
a given player statistic. Also provides the player's team.
}
\examples{
overallLeaders(var="Saves")
overallLeaders(var="Assists", n_leaders=15, games="conference", players="skaters", gender="men", verbose=FALSE)
}
