% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter Plot of Two Variables}
\usage{
scatterPlot(x, y, trend = FALSE, gender = "women")
}
\arguments{
\item{x}{variable on the x-axis of the scatterplot}

\item{y}{variable on the y-axis of the scatterplot}

\item{trend}{TRUE/FALSE, whether to display a smoothing trend line to the scatter plot}

\item{gender}{gender of data to graph, default to "women"}
}
\value{
a scatter plot of the two given variables
}
\description{
Scatter Plot of Two Variables
}
\examples{
scatterPlot("Goals", "Assists", TRUE)
scatterPlot("Goals", "Assists", FALSE, "men")
}
