% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGameStats.R
\name{scrapeGameStats}
\alias{scrapeGameStats}
\title{Scrape Game Statistics for one season of ECAC hockey games}
\usage{
scrapeGameStats(gender = "women", verbose = TRUE)
}
\arguments{
\item{gender}{'women' (default) or 'men'}

\item{verbose}{TRUE (default) for intermediate data scraping output messages, FALSE for no additional output}
}
\value{
dataframe of game statistics
}
\description{
Webscrape (using rvest) skater and goalie statistics for all ECAC games this season
}
\examples{
scrapeGameStats(verbose=FALSE)
scrapeGameStats(gender="men", verbose=FALSE)
}
