% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGameStatsByTeam.R
\name{scrapeGameStatsByTeam}
\alias{scrapeGameStatsByTeam}
\title{Scrape Game by Game stats by team}
\usage{
scrapeGameStatsByTeam(team, season = "2021-22", gender = "women")
}
\arguments{
\item{team}{chosen team to scrape}

\item{gender}{gender of data to scrape}
}
\value{
a dataframe of ecac data for a given team and gender
}
\description{
Scrape Game by Game stats by team
}
