% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeIndivStats.R
\name{scrapeIndivStats}
\alias{scrapeIndivStats}
\title{Scrape Individual Player Statistics}
\usage{
scrapeIndivStats(
  games = "all",
  players = "goalies",
  gender = "women",
  verbose = TRUE
)
}
\arguments{
\item{games}{collecting data for 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{players}{type of players stats to collect, 'goalies' (default) or 'skaters'}

\item{gender}{'women' (default) or 'men'}

\item{verbose}{TRUE (default) for intermediate data scraping output messages, FALSE for no additional output}
}
\value{
data frame of individual player statistics for all ECAC teams
}
\description{
Webscrape (using rvest) all individual player statistics for the current season
}
\examples{
scrapeIndivStats()
scrapeIndivStats(games="conference", players="skaters", gender="men", verbose=FALSE)
}
