% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeIndivStatsByTeam.R
\name{scrapeIndivStatsByTeam}
\alias{scrapeIndivStatsByTeam}
\title{Scrape Individual Player Statistics For One Team}
\usage{
scrapeIndivStatsByTeam(
  team,
  games = "all",
  players = "goalies",
  gender = "women"
)
}
\arguments{
\item{team}{ECAC team to scrape data for}

\item{games}{collecting data for 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{players}{type of players stats to collect, 'goalies' (default) or 'skaters'}

\item{gender}{'women' (default) or 'men'}
}
\value{
data frame of individual player statistics for one ECAC team
}
\description{
Webscrape (using rvest) all individual player statistics of a given team for the current season
}
\examples{
scrapeIndivStatsByTeam(team="brown")
scrapeIndivStatsByTeam(team="clarkson", games="conference", players="skaters", gender="men")
}
