% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{titleCase}
\alias{titleCase}
\title{Title Case a String (single word)}
\usage{
titleCase(string)
}
\arguments{
\item{string}{a single word}
}
\value{
the given string in title case
}
\description{
Title Case a String (single word)
}
\examples{
titleCase("foo") gives "Foo"
titleCase("sTRinG") gives "String"
}
