% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varIsValid}
\alias{varIsValid}
\title{Check Variable Name Validity}
\usage{
varIsValid(variableName, type)
}
\arguments{
\item{variableName}{name of Variable to check}

\item{type}{type of variable to check ('team', 'skaters', 'goalies', or' game')}
}
\value{
TRUE if variableName is valid, FALSE otherwise
}
\description{
Checks the validity of a given variable name for a certain type of ECAC hockey data
}
\examples{
varIsValid(variableName="Goals", type="skaters") # TRUE
varIsValid(varaibleName="Assists", type="goalies") # FALSE
}
