% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winLossChart.R
\name{winLossChart}
\alias{winLossChart}
\title{Produce a Win/Loss ggplot2 Bar Chart}
\usage{
winLossChart(team, gender = "women", byLocation = TRUE)
}
\arguments{
\item{team}{ECAC team to visualize}

\item{gender}{'women' (default) or 'men'}

\item{byLocation}{TRUE (default) to split Home/Away, FALSE otherwise}
}
\value{
ggplot2 bar chart object
}
\description{
Creates a ggplot2 bar chart object visualizing the wins and losses of a given ECAC team.
Can be faceted by home/away.
}
\examples{
winLossChart(team="brown")
winLossChart(team="clarkson", gender="men", byLocation=FALSE)
}
