% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genderIsValid}
\alias{genderIsValid}
\title{Checks Gender Formatting}
\usage{
genderIsValid(gender)
}
\arguments{
\item{gender}{gender to check}
}
\value{
TRUE if gender is formatted correctly, FALSE otherwise
}
\description{
Checks that the gender argument provided is formatted correctly for ecacr processing
}
\examples{
genderIsValid(gender="women") # TRUE
genderIsValid(gender="WOMEN") # FALSE
genderIsValid(gender="W") # FALSE
}
