% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopChart.R
\name{lollipopChart}
\alias{lollipopChart}
\title{Make a Lollipop Chart}
\usage{
lollipopChart(var, conf = FALSE, gender = "women")
}
\arguments{
\item{var}{variable to graph, must be given as a string}

\item{gender}{gender of data to graph}
}
\value{
returns a lollipop chart of the given variable, separated by team. Plots Games Played by default.
}
\description{
Make a Lollipop Chart
}
\examples{
lollipopChart()
lollipopChart("Goals")
lollipopChart("Goals", "men")
}
