% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopChart.R
\name{lollipopChart}
\alias{lollipopChart}
\title{Produce a ggplot2 Lollipop Chart}
\usage{
lollipopChart(var, games = "all", gender = "women")
}
\arguments{
\item{var}{summary statistic to display}

\item{games}{graphing 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{gender}{"women' (default) or 'men'}
}
\value{
a ggplot2 lollipop chart object
}
\description{
Create a ggplot2 lollipop chart (alternative to a bar chart) of overall season summary
statistics of all ECAC teams.
}
\examples{
lollipopChart(var="Goals")
lollipopChart(var="PenaltyMinutes", games="conference", gender="men")
}
