% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Produce a ggplot2 Scatter Plot}
\usage{
scatterPlot(x, y, games = "all", gender = "women", trend = FALSE)
}
\arguments{
\item{x}{x-axis variable}

\item{y}{y-axis variable}

\item{games}{graphing 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games}

\item{gender}{'women' (default) or 'men'}

\item{trend}{TRUE (default) for smoothing trend line, FALSE for no trend line}
}
\value{
a ggplot2 scatter plot object
}
\description{
Create a ggplot2 scatter plot object of summary team statistics. Can display a trend line
if desired.
}
\examples{
scatterPlot(x="Goals", y="Wins")
scatterPLot(x="Assists", y="PenaltyMinutes", games="conference", gender="men", trend=TRUE)
}
