% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGameStats.R
\name{scrapeGameStats}
\alias{scrapeGameStats}
\title{Scrape data for all ECAC games}
\usage{
scrapeGameStats(gender = "women", verbose = TRUE)
}
\arguments{
\item{gender}{the given gender to scrape games for}

\item{verbose}{TRUE for additional output, FALSE for no additional output}
}
\value{

}
\description{
Scrape data for all ECAC games
}
