% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeIndivStats.R
\name{scrapeIndivStats}
\alias{scrapeIndivStats}
\title{Scrape Individual Player Statistics (all ECAC Teams, one gender)}
\usage{
scrapeIndivStats(
  conf = FALSE,
  skaters = FALSE,
  gender = "women",
  verbose = TRUE
)
}
\arguments{
\item{conf}{TRUE for ecac data only, FALSE for conference and non-conference}

\item{skaters}{TRUE for skater data, FALSE for goalie data}

\item{gender}{"women" for women's data, "men" for men's data}

\item{verbose}{TRUE for extra output during runtime, FALSE for no extra output}
}
\value{
data frame of individual player statistics for all ECAC teams
}
\description{
Scrape Individual Player Statistics (all ECAC Teams, one gender)
}
