% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeIndivStatsByTeam.R
\name{scrapeIndivStatsByTeam}
\alias{scrapeIndivStatsByTeam}
\title{Scrape Individual Statistics (for one ECAC team)}
\usage{
scrapeIndivStatsByTeam(team, conf = FALSE, skaters = FALSE, gender = "women")
}
\arguments{
\item{team}{team of choice}

\item{conf}{TRUE for ecac data only, FALSE for conference and non-conference}

\item{skaters}{TRUE for skater data, FALSE for goalie data}

\item{gender}{"women" for women's data, "men" for men's data}
}
\value{
data frame of individual statistics of given parameters
}
\description{
Scrape Individual Statistics (for one ECAC team)
}
