% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeTeamStats.R
\name{scrapeTeamStats}
\alias{scrapeTeamStats}
\title{Scrape Team Statistics}
\usage{
scrapeTeamStats(gender = "women")
}
\value{
returns a either women's or men's team statistics data. Returns women's data by default.
}
\description{
Scrape Team Statistics
}
