% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teamIsValid}
\alias{teamIsValid}
\title{Check that a team name is valid}
\usage{
teamIsValid(team)
}
\arguments{
\item{team}{team name to check}
}
\value{
TRUE if team is in ECAC and formatted correctly, FALSE otherwise
}
\description{
Check that a team name is valid
}
