% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varIsValid}
\alias{varIsValid}
\title{Check Variable Name Validity}
\usage{
varIsValid(variableName)
}
\arguments{
\item{variableName}{name of Variable to check}
}
\value{
TRUE if variableName is valid, FALSE otherwise
}
\description{
Check Variable Name Validity
}
\examples{
varIsValid("Goals") returns TRUE,
varIsValid("notValid") returns FALSE
}
