% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varIsValid}
\alias{varIsValid}
\title{Check Team Variable Name Validity}
\usage{
varIsValid(variableName, team, skaters = NULL)
}
\arguments{
\item{variableName}{name of Variable to check}

\item{team}{TRUE if checking team variable names, FALSE otherwise}

\item{skaters}{TRUE if checking skater variable name, FALSE if checking goalie variable names, NULL otherwise}
}
\value{
TRUE if variableName is valid, FALSE otherwise
}
\description{
Check Team Variable Name Validity
}
