% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winLossChart.R
\name{winLossChart}
\alias{winLossChart}
\title{Produce a chart visualizing the wins and losses of a team}
\usage{
winLossChart(team, byLocation = TRUE, gender = "women")
}
\arguments{
\item{team}{the team you'd like to investigate}

\item{byLocation}{TRUE to split Home/Away, FALSE otherwise (default TRUE)}

\item{gender}{gender you'd like to investigate (default "women")}
}
\value{

}
\description{
Produce a chart visualizing the wins and losses of a team
}
